Create separate ms-codeview crate
