+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 
 members = [
+    "codeview",
     "msf",
     "msfz",
     "pdb",
