+++ codeview/src/syms/iter.rs
