+++ codeview/src/syms/offset_segment.rs
