+++ pdb/src/globals/psi.rs
@@ -9,11 +9,11 @@
 
 use super::gss::*;
 use super::name_table::*;
-use crate::parser::{Parse, Parser, ParserMut};
 use crate::syms::{OffsetSegment, Pub};
 use crate::utils::is_aligned_4;
 use anyhow::{bail, Context};
 use bstr::BStr;
+use ms_codeview::parser::{Parse, Parser, ParserMut};
 use std::mem::size_of;
 use tracing::{debug, error, info};
 use zerocopy::{FromBytes, Immutable, IntoBytes, KnownLayout, Unaligned, LE, U16, U32};
